Article views count support with Celery.
Adding a new field `views` to the Article model representing the total views an Article has. For better insights the views are not used
as a count but instead another model ArticleView is created. Everytime a user visits an Article, we insert the information into this table.

So here comes Celery into the picture. To keep track of the inserted rows for the views, we are using Celery to run our background task
to count the total number of views for each Article and then updating the Article's `views` field.

The task is scheduled to run on an interval of 5 minutes.

With celery we are using RabbitMQ as a message broker, flower to monitor the Celery cluster.

Updated Commands:

To run celery worker:
$ celery -A app.celery  worker -l info -B

To run flower:
$ celery -A app.celery flower

To run Flask application:
$ export FLASK_APP=app
$ flask run

Requirements.txt is updated to inlcude packages requried to run Celery with our application.